Backend

+++ urban_connection/app/Http/Controllers/TagsController.php
@@ -93,6 +93,10 @@ class TagsController extends Controller
      */
     public function destroy( Tag $tag)
     {
+        if($tag->posts->count()>0){
+            session()->flash('success', "Can't be deleted it has posts.");
+            return redirect()->back();
+        }
         $tag->delete();
         session()->flash('success', 'Deleted successfully.');
         return redirect(route('tags.index'));
